+++ Assets/Prefabs/CharDeTeste.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 1f8e5ee111129c7428baeebefc01c279
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
