+++ Assets/Scripts/Game/Characters/CharacterInventory.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f2e7feae723f69c40b5931e386d7b85c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
