+++ Assets/Scripts/Game/Objects.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 8ed1cad12baf44142bc8dba5414ddc69
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
