+++ Assets/Scripts/Game/System.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 89213eb2a9335914488c8980ac3ea266
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
