+++ Assets/Scripts/Game/System/GameManager.cs.meta
