+++ Assets/Scripts/Game/UI.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 42951d8742dde9e47bca65b6bc73118b
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
