+++ Assets/Scripts/Game/UI/PauseButton.cs
@@ -1,0 +1,11 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class PauseButton : MonoBehaviour {
+    
+    public void TogglePause() {
+        NetworkingPause.OwnedInstance.TogglePause();
+    }
+
+}
