--- Assets/Scripts/PauseButton.cs
@@ -1,11 +1,0 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class PauseButton : MonoBehaviour {
-    
-    public void TogglePause() {
-        NetworkingPause.OwnedInstance.TogglePause();
-    }
-
-}
