+++ ProjectSettings/TagManager.asset
@@ -15,7 +15,7 @@ TagManager:
   - 
   - Managers
   - Player
-  - 
+  - Interactables
   - 
   - 
   - 
@@ -40,4 +40,7 @@ TagManager:
   m_SortingLayers:
   - name: Default
     uniqueID: 0
+    locked: 0
+  - name: Interactables
+    uniqueID: 258518167
     locked: 0
